requirements.txt
openai-whisper==20250625
pyaudio==0.2.14
numpy==1.26.4
torch==2.11.0
ffmpeg-python==0.2.0
sounddevice==0.4.7
soundfile==0.12.1